--- src/Config/Database.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Config;
-
-class Database
-{
-    public static function loadEnv(): void
-    {
-        Config::init();
-    }
-}
